* origin/master: (34 commits)
  RTEMS 4.14 GCC patch 20110829
  overrides and source 3.0
  drop config.guess and config.sub
  4.4.6-1
  debianize
  newlib-1.18.0-rtems4.10-20110518.diff
  gcc-core-4.4.6-rtems4.10-20110420.diff
  4.4.5-2
  RTEMS 4.10 GGC patch 20110301
  RTEMS 4.10 newlib patch 20110114
  switch RTEMS version to 4.10
  debianize
  drop config.{guess,sub}
  RTEMS 4.10 newlib patch 20101006
  RTEMS 4.10 GGC patch 20101003
  4.3.2-3
  rtems newlib patch 20090324
  rtems gcc core patch 20090825
  Revert "rtems newlib patch 20081203"
  Revert "rtems gcc core patch 20081214"
  ...

+++ newlib/libc/posix/readdir.c
@@ -37,6 +37,7 @@
 static char sccsid[] = "@(#)readdir.c	5.7 (Berkeley) 6/1/90";
 #endif /* LIBC_SCCS and not lint */
 
+#include <stdint.h>
 #include <dirent.h>
 
 extern int getdents (int fd, void *dp, int count);
@@ -75,7 +76,7 @@ _DEFUN(readdir, (dirp),
       continue;
     }
     dp = (struct dirent *)(dirp->dd_buf + dirp->dd_loc);
-    if ((int)dp & 03) {	/* bogus pointer check */
+    if ((intptr_t)dp & 03) {	/* bogus pointer check */
 #ifdef HAVE_DD_LOCK
       __lock_release_recursive(dirp->dd_lock);
 #endif
